% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionInitialization.R
\name{open_window}
\alias{open_window}
\title{Open a new window}
\usage{
open_window()
}
\description{
Open new window in your web-browser (in case you accidentally closed the previous
one and don't want to return the \code{\link[start_session]{start_session()}})
}
\examples{
## Not run:
open_window()
# And... that's it
}
